# docs/swagger-admin-api.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FlowTim Admin Panel API",
    "description": "Admin panel API for managing users, workspaces, and platform settings",
    "version": "1.0.0",
    "contact": {
      "name": "FlowTim API Support",
      "email": "support@flowtim.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    },
    {
      "url": "https://api.flowtim.com",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/api/admin/settings": {
      "get": {
        "tags": ["Admin - Settings"],
        "summary": "Get system settings",
        "description": "Retrieve system configuration settings. Requires admin access.",
        "operationId": "getSettings",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Settings retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Settings fetched successfully"
                    },
                    "settings": {
                      "$ref": "#/components/schemas/SystemSettings"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          }
        }
      },
      "put": {
        "tags": ["Admin - Settings"],
        "summary": "Update system settings",
        "description": "Update system configuration settings. Requires admin access.",
        "operationId": "updateSettings",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemSettings"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Settings updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Settings updated successfully"
                    },
                    "settings": {
                      "$ref": "#/components/schemas/SystemSettings"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken",
        "description": "JWT access token in cookie"
      }
    },
    "schemas": {
...
      "SystemSettings": {
        "type": "object",
        "properties": {
          "maintenance": {
            "type": "object",
            "properties": {
              "enabled": {
                "type": "boolean",
                "example": false
              },
              "message": {
                "type": "string",
                "example": ""
              }
            }
          },
          "registration": {
            "type": "object",
            "properties": {
              "enabled": {
                "type": "boolean",
                "example": true
              },
              "requireEmailVerification": {
                "type": "boolean",
                "example": false
              }
            }
          },
          "features": {
            "type": "object",
            "properties": {
              "workspaces": {
                "type": "boolean",
                "example": true
              },
              "projects": {
                "type": "boolean",
                "example": true
              },
              "tasks": {
                "type": "boolean",
                "example": true
              },
              "chat": {
                "type": "boolean",
                "example": true
              },
              "documents": {
                "type": "boolean",
                "example": true
              }
            }
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "statusCode": {
            "type": "integer"
          },
          "errorCode": {
            "type": "string"
          }
        }
      }
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Authentication required - No valid token provided",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "message": "Authentication required",
              "statusCode": 401,
              "errorCode": "ACCESS_UNAUTHORIZED"
            }
          }
        }
      },
      "ForbiddenError": {
        "description": "Forbidden - Admin access required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "message": "Admin access required",
              "statusCode": 403,
              "errorCode": "ACCESS_UNAUTHORIZED"
            }
          }
        }
      },
...
    }
  }
}